Move wheel helper function to wheel module
